+++ src/main/java/racingcar/Application.java
@@ -3,6 +3,7 @@ package racingcar;
 public class Application {
 	public static void main(String[] args) {
 		// TODO 구현 진행
+		GameLogic.startGame();
 
 	}
 }
