+++ src/main/java/racingcar/check/NumberChecking.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.check;
 
 public class NumberChecking {
